requirements.txt
django
psycopg2-binary
supabase
gunicorn
whitenoise
django-extensions
python-decouple
django-cors-headers
djangorestframework-simplejwt
PyJWT<2.10.0
django-csp
django-axes
bleach>=6.0.0
python-magic
django-ckeditor 

